+++ src/components/AddJournal/AddJournal.js
@@ -27,38 +27,40 @@ import {
   ToggleContainer,
 } from './styles';
 import { FormInputBtn } from '../Authentication/styles';
+import { useGlobalContext } from '../../context/DataContext';
 
-const AddJournal = ({
-  title,
-  setTitle,
-  authors,
-  setAuthors,
-  journaltype,
-  setJournaltype,
-  topic,
-  setTopic,
-  issn,
-  setIssn,
-  link,
-  setLink,
-  policy,
-  setPolicy,
-  dataavail,
-  setDataavail,
-  handleChangeData,
-  datashared,
-  setDatashared,
-  handleChangeData2,
-  peerreview,
-  setPeerreview,
-  handleChangePeer,
-  enforced,
-  setEnforced,
-  evidence,
-  setEvidence,
-  isPending,
-  handleSubmit,
-}) => {
+const AddJournal = () => {
+  const {
+    title,
+    setTitle,
+    authors,
+    setAuthors,
+    journaltype,
+    setJournaltype,
+    topic,
+    setTopic,
+    issn,
+    setIssn,
+    link,
+    setLink,
+    policy,
+    setPolicy,
+    dataavail,
+    setDataavail,
+    handleChangeData,
+    datashared,
+    setDatashared,
+    handleChangeData2,
+    peerreview,
+    setPeerreview,
+    handleChangePeer,
+    enforced,
+    setEnforced,
+    evidence,
+    setEvidence,
+    isPending,
+    handleSubmit,
+  } = useGlobalContext();
   return (
     <Container>
       <PolicyContainer>
