+++ src/pages/Home.js
@@ -1,18 +1,20 @@
 import React from 'react';
 import { Header, LandingSection } from '../components';
 import '../components/Landing/Header/Header.css';
+import { useGlobalContext } from '../context/DataContext';
 
-function Home({
-  search,
-  setSearch,
-  searchResults,
-  posts,
-  filteredData,
-  setFilteredData,
-  wordEntered,
-  setWordEntered,
-  handleFilter,
-}) {
+function Home() {
+  const {
+    search,
+    searchResults,
+    setSearch,
+    posts,
+    filteredData,
+    setFilteredData,
+    wordEntered,
+    setWordEntered,
+    handleFilter,
+  } = useGlobalContext();
   return (
     <div className='header-padding'>
       <Header
