+++ src/pages/Journal/Journal.js
@@ -4,21 +4,24 @@
 import React from 'react';
 import { Journals } from '../../components';
 import { Container } from '../../components/Journals/styles';
+import { useGlobalContext } from '../../context/DataContext';
 
-const Journal = ({ posts, search, setSearch, loading, postsPerPage, totalPosts, paginate }) => {
+const Journal = () => {
+  const { posts, search, setSearch, loading, postsPerPage, totalPosts, paginate, currentPost } =
+    useGlobalContext();
   if (loading) {
     return <h2>loading...</h2>;
   }
   return (
     <Container>
-      {posts.length ? (
+      {currentPost.length ? (
         <Journals
-          posts={posts}
+          posts={currentPost}
           search={search}
           setSearch={setSearch}
           loading={loading}
           postsPerPage={postsPerPage}
-          totalPosts={totalPosts}
+          totalPosts={posts.length}
           paginate={paginate}
         />
       ) : (
